Re #341: Prevent indexing of harvested Europeana records if saving to db fails.

+++ app/jobs/europeana_harvest_job.rb
@@ -63,15 +63,16 @@ private
   def create_record(record_id)
     record = EuropeanaRecord.find_or_initialize_by_record_id(record_id)
     if record.new_record?
-      record.save
-      begin
-        record.index
-        Sunspot.commit
-      rescue
-        # Indexing failed; do not store this record
-        RunCoCo.error_logger.error("#{self.class.to_s} failed to index Europeana record with ID \"#{record_id}\"")
-        record.destroy
-        return
+      if record.save
+        begin
+          record.index
+          Sunspot.commit
+        rescue
+          # Indexing failed; do not store this record
+          RunCoCo.error_logger.error("#{self.class.to_s} failed to index Europeana record with ID \"#{record_id}\"")
+          record.destroy
+          return
+        end
       end
       @harvested = @harvested + 1
     end
